requirements.txt
# Legal RAG System Dependencies
# Install with: pip install -r requirements.txt

# Document Processing
pymupdf4llm>=0.0.10
docling>=2.0.0           # Optional: high-quality PDF (needs >2GB RAM)
surya-ocr>=0.17.0        # OCR for scanned/image-based PDFs (Greek + 90 languages)
psutil>=5.9.0            # RAM detection for Docling auto-enable

# Embeddings & Reranking
cohere>=5.0.0
voyageai>=0.2.0

# LLM Client (NVIDIA NIM API via OpenAI-compatible client)
openai>=1.0.0

# Vector Database
psycopg2-binary>=2.9.0
pgvector>=0.3.0

# Web Interface
streamlit>=1.30.0

# API Backend
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
